third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-text/white-space/seg-break-transformation-017-expected.txt

This is a testharness.js-based test.
FAIL linebreak only assert_equals: expected 98 but got 105
PASS zwsp retained 1
FAIL spaces linebreak assert_equals: expected 98 but got 105
PASS zwsp retained 2
FAIL linebreak spaces assert_equals: expected 98 but got 105
PASS zwsp retained 3
FAIL spaces linebreak spaces assert_equals: expected 98 but got 105
PASS zwsp retained 4
FAIL multiple linebreaks assert_equals: expected 98 but got 105
PASS zwsp retained 5
FAIL multiple linebreaks + spaces assert_equals: expected 98 but got 105
PASS zwsp retained 6
Harness: the test ran to completion.
